+++ src/Sentry.Maui/Internal/IMauiEventsBinder.cs
@@ -1,0 +1,6 @@
+namespace Sentry.Maui.Internal;
+
+internal interface IMauiEventsBinder
+{
+    void BindMauiEvents();
+}