+++ src/Sentry/Internal/Enricher.cs
@@ -17,6 +17,7 @@ namespace Sentry.Internal
             {
                 Name = current.Name,
                 Version = current.Version,
+                Identifier = current.Identifier,
                 RawDescription = current.Raw
             };
         });
@@ -54,6 +55,9 @@ namespace Sentry.Internal
             {
                 eventLike.Sdk.AddPackage("nuget:" + SdkVersion.Instance.Name, SdkVersion.Instance.Version);
             }
+
+            // Platform
+            eventLike.Platform ??= Sentry.Constants.Platform;
 
             // Release
             eventLike.Release ??= ReleaseLocator.Resolve(_options);