+++ test/Sentry.Maui.Tests/SentryMauiAppBuilderExtensionsTests.cs
@@ -1,16 +1,29 @@
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using Microsoft.Maui.Hosting;
+using MauiConstants = Sentry.Maui.Internal.Constants;
 
 namespace Sentry.Maui.Tests;
 
 public class SentryMauiAppBuilderExtensionsTests
 {
+    private class Fixture
+    {
+        public MauiAppBuilder Builder { get; }
+
+        public Fixture()
+        {
+            var builder = MauiApp.CreateBuilder();
+            builder.Services.AddSingleton(Substitute.For<IApplication>());
+            Builder = builder;
+        }
+    }
+
+    private readonly Fixture _fixture = new();
+
     [Fact]
     public void CanUseSentry_WithConfigurationOnly()
     {
         // Arrange
-        var builder = MauiApp.CreateBuilder();
+        var builder = _fixture.Builder;
         builder.Services.Configure<SentryMauiOptions>(options =>
         {
             options.Dsn = DsnSamples.ValidDsnWithoutSecret;
@@ -33,7 +46,7 @@ public class SentryMauiAppBuilderExtensionsTests
     public void CanUseSentry_WithDsnStringOnly()
     {
         // Arrange
-        var builder = MauiApp.CreateBuilder();
+        var builder = _fixture.Builder;
 
         // Act
         var chainedBuilder = builder.UseSentry(DsnSamples.ValidDsnWithoutSecret);
@@ -52,7 +65,7 @@ public class SentryMauiAppBuilderExtensionsTests
     public void CanUseSentry_WithOptionsOnly()
     {
         // Arrange
-        var builder = MauiApp.CreateBuilder();
+        var builder = _fixture.Builder;
 
         // Act
         var chainedBuilder = builder.UseSentry(options =>
@@ -74,7 +87,7 @@ public class SentryMauiAppBuilderExtensionsTests
     public void CanUseSentry_WithConfigurationAndOptions()
     {
         // Arrange
-        var builder = MauiApp.CreateBuilder();
+        var builder = _fixture.Builder;
         builder.Services.Configure<SentryMauiOptions>(options =>
         {
             options.Dsn = DsnSamples.ValidDsnWithoutSecret;
@@ -100,7 +113,7 @@ public class SentryMauiAppBuilderExtensionsTests
     public void UseSentry_EnablesGlobalMode()
     {
         // Arrange
-        var builder = MauiApp.CreateBuilder();
+        var builder = _fixture.Builder;
 
         // Act
         _ = builder.UseSentry(DsnSamples.ValidDsnWithoutSecret);
@@ -117,7 +130,7 @@ public class SentryMauiAppBuilderExtensionsTests
     {
         // Arrange
         SentryEvent @event = null;
-        var builder = MauiApp.CreateBuilder()
+        var builder = _fixture.Builder
             .UseSentry(options =>
             {
                 options.Dsn = DsnSamples.ValidDsnWithoutSecret;
@@ -138,7 +151,45 @@ public class SentryMauiAppBuilderExtensionsTests
 
         // Assert
         Assert.NotNull(@event);
-        Assert.Equal(Constants.SdkName, @event.Sdk.Name);
-        Assert.Equal(Constants.SdkVersion, @event.Sdk.Version);
+        Assert.Equal(MauiConstants.SdkName, @event.Sdk.Name);
+        Assert.Equal(MauiConstants.SdkVersion, @event.Sdk.Version);
+    }
+
+    [Fact]
+    public void UseSentry_EnablesHub()
+    {
+        // Arrange
+        var builder = _fixture.Builder
+            .UseSentry(DsnSamples.ValidDsnWithoutSecret);
+
+        // Act
+        using var app = builder.Build();
+        var hub = app.Services.GetRequiredService<IHub>();
+
+        // Assert
+        Assert.True(hub.IsEnabled);
+    }
+
+    [Fact]
+    public void UseSentry_AppDispose_DisposesHub()
+    {
+        // Note: It's crucial to dispose the hub when the app disposes, so that we flush events from the
+        //       queue in the BackgroundWorker.  The app container will dispose any services registered
+        //       that implement IDisposable.
+
+        // Arrange
+        var builder = _fixture.Builder
+            .UseSentry(DsnSamples.ValidDsnWithoutSecret);
+
+        // Act
+        IHub hub;
+        using (var app = builder.Build())
+        {
+            hub = app.Services.GetRequiredService<IHub>();
+        }
+
+        // Assert
+        // Note, the hub is disabled when disposed.  We ensure it's first enabled in the previous test.
+        Assert.False(hub.IsEnabled);
     }
 }